pyproject.toml
[project]
name = "zibal-backend"
version = "1.0.0"
description = "Zibal Backend Interview Project"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "Django>=4.2.0",
    "djangorestframework>=3.14.0",
    "mongoengine>=0.27.0",
    "pymongo>=4.3.3",
    "celery>=5.3.0",
    "jdatetime>=4.1.0",
    "uvicorn>=0.24.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]



[tool.black]
color = true
line-length = 120
target-version = ['py311']
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.ruff_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff.per-file-ignores]
"features/steps/*.py" = ["F811"] # Just for redefinition of unused gherkin step and imp function

[tool.ruff]
fix = false
select = [
    "B", # flake8-bugbear
    "C", # flake8-comprehensions
    "C4", # comprehensions
    "E", # pycodestyle errors
    "F", # pyflakes
    "I", # isort
    "INT", # flake8-gettext
    "ISC", # implicit string concatenation
    "PGH", # pygrep-hooks
    "PIE", # misc lints
    "PLC", "PLE", "PLR", "PLW", # pylint
    "PYI", # flake8-pyi
    "Q", # flake8-quotes
    "RUF", # Ruff-specific rules
    "T10", # flake8-debugger
    "TCH", # type-checking imports
    "TID", # tidy imports
    "W", # pycodestyle warnings
    "YTT", # flake8-2020
]
ignore = [
    "C901", # too complex
    "RUF001",
    "RUF003", # String contains ambiguous
    "PLR2004", # Magic value used in comparison
    "PLR0912",
    "PLR0912",
    "PLR0915",
    "PLR0911",
    "PLR0912",
    "Q000",
    "PLR0913",
    "F405",
    "W191",
]
src = ["src"]

line-length = 120
target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.pylint]
max-args = 5
max-branches = 12
max-returns = 6
max-statements = 50

[tool.autoflake]
check = false
in-place = true
recursive = true
remove-all-unused-imports = true
ignore-init-module-imports = true

[tool.ruff.isort]
combine-as-imports = true
known-first-party = ["src"]
section-order = ["standard-library", "third-party", "first-party", "local-folder", ]

[tool.isort]
combine_as_imports = true
ensure_newline_before_comments = true
filter_files = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
profile = "black"
py_version = 311
use_parentheses = true

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = false
disallow_untyped_decorators = true
disallow_untyped_defs = true
follow_imports = 'silent'
no_implicit_reexport = true
pretty = true
python_version = "3.11"
show_error_codes = true
show_error_context = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.bandit]
exclude_dirs = ["tests", "features"]
skips = ["B101", "B403", "B301"]